#268 moving some of the my location overlays into the correct package and adding setters for bitmap images #127 relate issue with setting the location icon
